Fix error on antelope-rs serialization branch that caused the error while trying to get addresses given an account, also add better error on 1559s and update test container to use the one without block_delta

+++ src/types/evm_types.rs
@@ -73,7 +73,7 @@ pub struct WithdrawAction {
 pub struct PrintedReceipt {
     pub charged_gas: String,
     pub trx_index: u16,
-    pub block: u32,
+    pub block: u64,
     pub status: u8,
     pub epoch: u64,
     pub createdaddr: String,
